Error handling + code adjustment

+++ src/main/java/com/onebox/test/OneBoxTest/controllers/ProductController.java
@@ -33,7 +33,7 @@ public class ProductController {
 	}
 	
 	@GetMapping(path="/{productId}")
-	public Optional<ProductModel> getProductById(@PathVariable int productId){
+	public Optional<ProductModel> getProductById(@PathVariable Long productId){
 		return this.productService.getProductById(productId);
 	}
